requirements.txt
# Core serving
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4

# LLM + fine-tuned model
torch==2.3.1          # CPU or CUDA — Docker will handle CUDA variant
transformers==4.51.3
peft==0.15.2
bitsandbytes==0.43.3  # CPU-compatible for Docker; swap 0.49.2 if CUDA 12.8
accelerate==1.6.0

# RAG
langchain==0.3.25
langchain-community==0.3.23
pinecone-client==5.0.1
sentence-transformers==4.1.0

# Utilities
python-dotenv==1.0.1